Move validator to common
