+++ Src/DeUrgenta.Common/DeUrgenta.Common.csproj
@@ -8,6 +8,7 @@
 
 
 	<ItemGroup>
+		<PackageReference Include="MediatR" Version="9.0.0" />
 		<PackageReference Include="Microsoft.EntityFrameworkCore" Version="5.0.4" />
 		<PackageReference Include="Swashbuckle.AspNetCore.Filters" Version="6.1.0" />
 	</ItemGroup>
