+++ Src/DeUrgenta.Common/Validation/IValidateRequest.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading.Tasks;
 using MediatR;
 
-namespace DeUrgenta.Group.Api.Validators
+namespace DeUrgenta.Common.Validation
 {
     public interface IValidateRequest<in T> where T : IBaseRequest
     {
