+++ Src/DeUrgenta.Group.Api/QueryHandlers/GetMyGroupsHandler.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using CSharpFunctionalExtensions;
+using DeUrgenta.Common.Validation;
 using DeUrgenta.Domain;
 using DeUrgenta.Group.Api.Models;
 using DeUrgenta.Group.Api.Queries;
