+++ Src/DeUrgenta.Group.Api/Validators/GetGroupSafeLocationsValidator.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Threading.Tasks;
+using DeUrgenta.Common.Validation;
 using DeUrgenta.Domain;
 using DeUrgenta.Group.Api.Queries;
 using Microsoft.EntityFrameworkCore;
