+++ Src/Tests/DeUrgenta.Group.Api.Tests/QueriesHandlers/GetAdministeredGroupsHandlerShould.cs
@@ -1,5 +1,6 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
+using DeUrgenta.Common.Validation;
 using DeUrgenta.Domain;
 using DeUrgenta.Group.Api.Queries;
 using DeUrgenta.Group.Api.QueryHandlers;
